// src/app/shared/models/pokemon.model.ts
export interface Pokemon {

    abilities: Ability2[];
    baseExperience: number;
    sprites: Sprites;
    name: string;
    stats: Stat[];
}
// ...
interface Stat {
  base_stat: number;
  effort: number;
  stat: Ability;
}

interface Sprites {
  back_default: string;
  back_female?: any;
  back_shiny: string;
  back_shiny_female?: any;
  front_default: string;
  front_female?: any;
  front_shiny: string;
  front_shiny_female?: any;
  other: Other;
  versions: Versions;
}

interface Versions {
  'generation-i': Generationi;
  'generation-ii': Generationii;
  'generation-iii': Generationiii;
  'generation-iv': Generationiv;
  'generation-v': Generationv;
  'generation-vi': Generationvi;
  'generation-vii': Generationvii;
  'generation-viii': Generationviii;
}
// ...
interface Other {
  dream_world: Dreamworld;
  home: Home;
  'official-artwork': Officialartwork;
}
// ...
interface Ability2 {
  ability: Ability;
  is_hidden: boolean;
  slot: number;
}

interface Ability {
  name: string;
  url: string;
}